Integrate new endpoints for updating and getting website URL
Integrate new endpoints for updating and retrieving the website URL into the `endpoints.py` file.

* Add import statement for `WebsiteUpdateRequest` and `current_website` from `update_website.py`.
* Add new endpoint `/update_website/` to update the website URL.
* Add new endpoint `/current_website/` to get the current website URL.
* Include the new router in the FastAPI application.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/akaday/py-sky-wifi?shareId=XXXX-XXXX-XXXX-XXXX).
